requirements.txt
beautifulsoup4
requests
urllib3
selenium
python-dotenv